* Changed some paremeters and return values to use size_t and ssize_t instead of int.
git-svn-id: svn://cybergarage.org:3960/project/uhttp/trunk@54 7ed65ffe-41cb-4a7d-888d-b4931f95f61e

+++ include/uhttp/io/BufferedReader.h
@@ -29,7 +29,7 @@ public:
 		this->reader = reader;
 	}
 
-	int read(std::string &b, int len)
+	ssize_t read(std::string &b, size_t len)
 	{
 		return reader->read(b, len);
 	}
@@ -39,12 +39,12 @@ public:
 		return reader->skip(n);
 	}
 
-	void unread(std::string &b, int off, int len)
+	void unread(std::string &b, size_t off, size_t len)
 	{
 		reader->unread(b, off, len);
 	}
 
-	void unread(std::string &b, int len)
+	void unread(std::string &b, size_t len)
 	{
 		reader->unread(b, len);
 	}
