+++ include/uhttp/io/FileInputStream.h
@@ -31,10 +31,10 @@ public:
 	FileInputStream(File *file, const std::string &mode);
 	~FileInputStream();
 
-	int read(std::string &b, int len);
-	int read(char *b, int len);
+	ssize_t read(std::string &b, size_t len);
+	ssize_t read(char *b, size_t len);
 
-	void unread(std::string &b, int off, int len);  // Not Implemented
+	void unread(std::string &b, size_t off, size_t len);  // Not Implemented
 
 	long skip(long n);
