+++ include/uhttp/io/InputStream.h
@@ -29,17 +29,17 @@ public:
 	{
 	}
 
-	virtual int read(std::string &b, int len) = 0;
-	virtual int read(char *b, int len) = 0;
+	virtual ssize_t read(std::string &b, size_t len) = 0;
+	virtual ssize_t read(char *b, size_t len) = 0;
 
-	int read(unsigned char *b, int len)
+	ssize_t read(unsigned char *b, size_t len)
 	{
 		return read((char *)b, len);
 	}
 
-	virtual void unread(std::string &b, int off, int len) = 0;  
+	virtual void unread(std::string &b, size_t off, size_t len) = 0;
 
-	void unread(std::string &b, int len)
+	void unread(std::string &b, size_t len)
 	{
 		 unread(b, 0, len);  
 	}
